@@ -111,12 +111,16 @@
 		<ul>
 			{#each tabs as tab}
 				{#if !tab.nest}
-					<li><div class="navLink" on:click={() => goto(tab.dest)}>{tab.label}</div></li>
+					<li><div class="navLink" onclick={() => goto(tab.dest)}>{tab.label}</div></li>
 				{:else}
 					{#each tab.children as child}
                         <!-- Shouldn't show Managers tab unless managers has been populated -->
 				        {#if child.label != "Managers" || managers.length > 0}
-                            <li><div class="navLink" on:click={() => goto(child.dest)}>{child.label}</div></li>
+							{#if child.label == "Go to Sleeper"}
+								<li><div class="navLink" onclick={() => window.location = child.dest}>{child.label}</div></li>
+							{:else}
+                            	<li><div class="navLink" onclick={() => goto(child.dest)}>{child.label}</div></li>
+							{/if}
                         {/if}
 					{/each}
 				{/if}
