@@ -198,6 +198,7 @@
 	}
 
 	.playerAvatar {
+        position: relative;
 		vertical-align: middle;
 		height: 45px;
 		width: 45px;
@@ -455,10 +456,32 @@
         display: flex;
         align-items: center;
     }
+    .teamLogo {
+        width: 21px;
+        position: absolute;
+        top: 0;
+    }
+    .teamHomeLogo {
+        right: -16px;
+    }
+    .teamAwayLogo {
+        left: -16px;
+    }
+    @media (max-width: 340px) {
+        .teamLogo {
+            width: 20px;
+        }
+        .teamHomeLogo {
+            right: -7px;
+        }
+        .teamAwayLogo {
+            left: -7px;
+        }
+    }
 </style>
 
 <div class="matchup">
-    <div class="header" on:click={() => expandClose()} bind:this={el} >
+    <div class="header" onclick={() => expandClose()} bind:this={el} >
         <div class="opponent home{winning == "home" ? " homeGlow" : ""}">
             <img class="avatar" src={home.manager.avatar} alt="home team avatar" />
             <div class="name">{home.manager.name}</div>
@@ -481,7 +504,11 @@
                             <span class="pos {player.home.pos}">{player.home.pos}</span>
                         {/if}
                         {#if player.home.avatar}
-                            <div class="playerAvatar playerInfo" style="{player.home.avatar}" />
+                            <div class="playerAvatar playerInfo" style="{player.home.avatar}">
+                                {#if player.home.team && player.home.pos != "DEF"}
+                                    <img src="https://sleepercdn.com/images/team_logos/nfl/{player.home.team.toLowerCase()}.png" class="teamLogo teamHomeLogo" alt="team logo"/>
+                                {/if}
+                            </div>
                         {/if}
                     </span>
                     <div class="nameHolder nameHolderL{player.home.name == 'Empty'? ' playerEmpty' : ''}">
@@ -502,7 +529,11 @@
                 <div class="player playerAway">
                     <span class="iconAndTeam iconAndTeamAway">
                         {#if player.away.avatar}
-                            <div class="playerAvatar playerInfo" style="{player.away.avatar}" />
+                            <div class="playerAvatar playerInfo" style="{player.away.avatar}">
+                                {#if player.away.team && player.away.pos != "DEF"}
+                                    <img src="https://sleepercdn.com/images/team_logos/nfl/{player.away.team.toLowerCase()}.png" class="teamLogo teamAwayLogo" alt="team logo"/>
+                                {/if}
+                            </div>
                         {/if}
                         {#if player.away.pos}
                             <span class="pos {player.away.pos}">{player.away.pos}</span>
@@ -523,7 +554,7 @@
             </div>
         {/each}
         {#if !expandOverride}
-            <div class="close" on:click={() => expandClose()}>Close Matchup</div>
+            <div class="close" onclick={() => expandClose()}>Close Matchup</div>
         {/if}
     </div>
 </div>
