@@ -1,18 +1,22 @@
 <script>
+	import { tabs } from '$lib/utils/tabs';
 	import Tab, { Icon, Label } from '@smui/tab';
 	import List, { Item, Graphic, Text, Separator } from '@smui/list';
 	import TabBar from '@smui/tab-bar';
+    import { page } from '$app/state';
 	import { goto, preloadData } from '$app/navigation';
 	import { enableBlog, managers } from '$lib/utils/leagueInfo';
 
-	export let active, tabs;
+	let active = $state(tabs.find(tab => tab.dest == page.url.pathname || (tab.nest && tab.children.find(subTab => subTab.dest == page.url.pathname))));
 
-	let activeTab = active;
+	let display = $state(false);
+	let el = $state();
+	let width = $state();
+	let height= $state();
+	let left = $state();
+	let top = $state();
 
-	let display;
-	let el, width, height, left, top;
-
-	const sizeSubMenu = (w) => {
+	$effect(() => {
 		top = el?.getBoundingClientRect() ? el?.getBoundingClientRect().top  : 0;
 		const bottom = el?.getBoundingClientRect() ? el?.getBoundingClientRect().bottom  : 0;
 
@@ -22,20 +26,11 @@
 		const right = el?.getBoundingClientRect() ? el?.getBoundingClientRect().right  : 0;
 
 		width = right - left;
-	}
+	});
 
-	let innerWidth;
+	let innerWidth = $state();
 
-	$: sizeSubMenu(innerWidth);
-
-	const open = (close = false) => {
-		if(close) {
-			setTimeout(() => {
-				active = activeTab;
-			}, 500)
-		} else {
-			activeTab = active;
-		}
+	const open = () => {
 		display = !display;
 	}
 
@@ -44,7 +39,7 @@
 		goto(dest);
 	}
 
-	let tabChildren = []
+	let tabChildren = $state([]);
 
 	for(const tab of tabs) {
 		if(tab.nest) {
@@ -106,40 +101,42 @@
 	}
 </style>
 
-<div class="overlay" style="display: {display ? "block" : "none"};" on:click={() => open(true)} />
+<div tabindex="0" role="button" class="overlay" style="display: {display ? "block" : "none"};" onclick={() => open(true)}></div>
 
 <div class="parent">
-	<TabBar class="navBar" {tabs} let:tab bind:active>
-		{#if tab.nest}
-			<div bind:this={el}>
+	<TabBar class="navBar" {tabs} key={(tab) => tab.key} bind:active>
+		{#snippet tab(tab)}
+			{#if tab.nest}
+				<div bind:this={el}>
+					<Tab
+						{tab}
+						minWidth
+						onclick={() => open()}
+					>
+						<Icon class="material-icons">{tab.icon}</Icon>
+						<Label>{tab.label}</Label>
+					</Tab>
+				</div>
+			{:else}
 				<Tab
+					class="{tab.label == 'Blog' && !enableBlog ? 'dontDisplay' : ''}"
 					{tab}
-					on:click={() => open(display)}
+					onTouchstart={() => preloadData(tab.dest)}
+					onMouseover={() => preloadData(tab.dest)}
+					href={tab.dest}
 					minWidth
 				>
 					<Icon class="material-icons">{tab.icon}</Icon>
 					<Label>{tab.label}</Label>
 				</Tab>
-			</div>
-		{:else}
-			<Tab
-				class="{tab.label == 'Blog' && !enableBlog ? 'dontDisplay' : ''}"
-				{tab}
-				on:touchstart={() => preloadData(tab.dest)}
-				on:mouseover={() => preloadData(tab.dest)}
-				on:click={() => goto(tab.dest)}
-				minWidth
-			>
-				<Icon class="material-icons">{tab.icon}</Icon>
-				<Label>{tab.label}</Label>
-			</Tab>
-		{/if}
+			{/if}
+		{/snippet}
 	</TabBar>
 	<div class="subMenu" style="max-height: {display ? 49 * tabChildren.length - 1 - (managers.length ? 0 : 48) : 0}px; width: {width}px; top: {height}px; left: {left}px; box-shadow: 0 0 {display ? "3px" : "0"} 0 #00316b; border: {display ? "1px" : "0"} solid #00316b; border-top: none;">
 		<List>
 			{#each tabChildren as subTab, ix}
 				{#if subTab.label == 'Managers'}
-					<Item class="{managers.length ? '' : 'dontDisplay'}" on:SMUI:action={() => subGoto(subTab.dest)} on:touchstart={() => preloadData(subTab.dest)} on:mouseover={() => preloadData(subTab.dest)}>
+					<Item class="{managers.length ? '' : 'dontDisplay'}" onSMUIAction={() => subGoto(subTab.dest)} ontouchstart={() => preloadData(subTab.dest)} onmouseover={() => preloadData(subTab.dest)}>
 						<Graphic class="material-icons">{subTab.icon}</Graphic>
 						<Text class="subText">{subTab.label}</Text>
 					</Item>
@@ -147,7 +144,7 @@
 						<Separator />
 					{/if}
 				{:else}
-					<Item on:SMUI:action={() => subGoto(subTab.dest)} on:touchstart={() => {if(subTab.label != 'Go to Sleeper') preloadData(subTab.dest)}} on:mouseover={() => {if(subTab.label != 'Go to Sleeper') preloadData(subTab.dest)}}>
+					<Item onSMUIAction={() => subGoto(subTab.dest)} ontouchstart={() => {if(subTab.label != 'Go to Sleeper') preloadData(subTab.dest)}} onmouseover={() => {if(subTab.label != 'Go to Sleeper') preloadData(subTab.dest)}}>
 						<Graphic class="material-icons">{subTab.icon}</Graphic>
 						<Text class="subText">{subTab.label}</Text>
 					</Item>
