@@ -55,7 +55,8 @@
 					break;
 			}
 			player = {
-				name: `${passedPlayers[singlePlayer].fn} ${passedPlayers[singlePlayer].ln}${injury ? `<span class="injury ${injury}">${injury}</span>` : ""}${roster.metadata && roster.metadata[`p_nick_${singlePlayer}`] ? `<br /><span class="nickname">"${roster.metadata[`p_nick_${singlePlayer}`]}"</span>` : ""}`,
+				name: `${passedPlayers[singlePlayer].fn} ${passedPlayers[singlePlayer].ln}${injury ? `<span class="injury ${injury}">${injury}</span>` : ""}`,
+                nickname: roster.metadata && roster.metadata[`p_nick_${singlePlayer}`] ? roster.metadata[`p_nick_${singlePlayer}`] : null,
 				poss: passedPlayers[singlePlayer].pos,
 				team: passedPlayers[singlePlayer].t,
 				avatar: passedPlayers[singlePlayer].pos == "DEF" ? `background-image: url(https://sleepercdn.com/images/team_logos/nfl/${singlePlayer.toLowerCase()}.png)` : `background-image: url(https://sleepercdn.com/content/nfl/players/thumb/${singlePlayer}.jpg), url(https://sleepercdn.com/images/v2/icons/player_default.webp)`,
@@ -262,14 +263,14 @@
 		<Head> <!-- Team name  -->
 			<Row>
 				<Cell colspan=4 class="r_{division} clickable">
-					<h3 on:click={() => gotoManager({leagueTeamManagers, rosterID: roster.roster_id})}>
+					<h3 onclick={() => gotoManager({leagueTeamManagers, rosterID: roster.roster_id})}>
 						<img alt="team avatar" class="teamAvatar" src="{team ? team.avatar : 'https://sleepercdn.com/images/v2/icons/player_default.webp'}" />
 						{team?.name ? team.name : 'No Manager'}
 					</h3>
 
 					<div class="record">
 						{#each record as result}
-							<img alt="match result" class="result" src="./{result}.png" />
+							<img alt="match result" class="result" src="/{result}.png" />
 						{/each}
 					</div>
 				</Cell>
@@ -280,7 +281,7 @@
 			{#each finalStarters as starter}
 				<RosterRow player={starter} />
 			{/each}
-			<Row class="interactive" on:click={toggleSelected}>
+			<Row class="interactive" onclick={toggleSelected}>
 				<Cell colspan=4 class="{division}"><h5><Icon class="material-icons icon">king_bed</Icon> Bench <span class="italic">({status})</span></h5></Cell>
 			</Row>
 		</Body>
@@ -302,7 +303,7 @@
 						<RosterRow player={ir} />
 					{/each}
 				{/if}
-				<Row class="interactive" on:click={toggleSelected}>
+				<Row class="interactive" onclick={toggleSelected}>
 					<Cell colspan=4 class="{division}"><h5><Icon class="material-icons icon">close_fullscreen</Icon>Close Bench</h5></Cell>
 				</Row>
 			</Body>
