@@ -8,6 +8,7 @@
 
     let post;
     let createdAt;
+    let id;
     let loading = true;
     let leagueTeamManagers = {};
 
@@ -18,6 +19,7 @@
             if(singlePost.fields.featured) {
                 createdAt = singlePost.sys.createdAt;
                 post = singlePost.fields;
+                id = singlePost.sys.id;
                 break;
             }
         }
@@ -28,6 +30,7 @@
                 if(singlePost.fields.featured) {
                     createdAt = singlePost.sys.createdAt;
                     post = singlePost.fields;
+                    id = singlePost.sys.id;
                     break;
                 }
             }
@@ -77,7 +80,7 @@
     </div>
 {:else}
     <h2>League Blog</h2>
-    <Post {leagueTeamManagers} {post} {createdAt} home={true} />
+    <Post {leagueTeamManagers} {post} {createdAt} {id} />
     <div class="center">
         <a class="viewAll" href="/blog">View More Blog Posts</a>
     </div>
