@@ -1,22 +1,9 @@
 <script>
-    import { parseDate, getAuthor, getAvatar, generateParagraph } from "$lib/utils/helper";
-    import { onMount } from "svelte";
+    import { generateParagraph } from "$lib/utils/helper";
     import { fly } from "svelte/transition";
-    import Comments from "./Comments.svelte";
+	import AuthorAndDate from "./AuthorAndDate.svelte";
 
-    export let leagueTeamManagers, post, createdAt, id = null, direction = 1, home = false;
-
-    let loadingComments = true;
-    let total, comments;
-
-    onMount(async()=> {
-        const res = await fetch(`/api/getBlogComments/${id}`, {compress: true});
-        const commentsData = await res.json();
-
-        total = commentsData.total;
-        comments = [...commentsData.items].sort((a, b) => Date.parse(a.sys.createdAt) - Date.parse(b.sys.createdAt));
-        loadingComments = false;
-    });
+    export let leagueTeamManagers, post, createdAt, id = null, direction = 1;
 
     let safePost = false;
     let title, body, type, author;
@@ -37,6 +24,10 @@
     }
 
     const duration = 300;
+
+    let e;
+
+    $: isOverflown = e ? e.scrollHeight > e.clientHeight : false;
 </script>
 
 <style>
@@ -55,82 +46,73 @@
         margin: 0;
     }
 
-    .authorAndDate a {
-        background-color: #00316b;
-        color: #fff;
+    .button {
+        background-color: #0082c3;
+        font-size: 1em;
         border-radius: 1em;
         text-decoration: none;
-        font-size: 0.8em;
         padding: 0.5em 1em;
         margin-right: 1em;
+        color: white;
     }
 
-    .authorAndDate a:hover {
-        background-color: #0082c3;
+    .button:hover {
+        background-color: #00316b;
     }
 
     :global(.body blockquote) {
-        margin: 0.1em 0;
         border-left: 3px solid rgb(231, 235, 238);
         margin: 1em 2em;
         padding-left: 0.875em;
     }
 
     :global(.body .heading-1) {
-        margin: 0.4em 0;
-        padding: 0 2em;
+        padding: 0.4em 2em;
         font-size: 1.9em;
         text-align: center;
     }
 
     :global(.body .heading-2) {
-        margin: 0.4em 0;
-        padding: 0 2em;
+        padding: 0.4em 2em;
         font-size: 1.8em;
         text-align: center;
     }
 
     :global(.body .heading-3) {
-        margin: 0.4em 0;
-        padding: 0 2em;
+        padding: 0.4em 2em;
         font-size: 1.7em;
         text-align: center;
     }
 
     :global(.body .heading-4) {
-        margin: 0.4em 0;
-        padding: 0 2em;
+        padding: 0.4em 2em;
         font-size: 1.6em;
         text-align: center;
     }
 
     :global(.body .heading-5) {
-        margin: 0.4em 0;
-        padding: 0 2em;
+        padding: 0.4em 2em;
         font-size: 1.5em;
         text-align: center;
     }
 
     :global(.body .heading-6) {
-        margin: 0.4em 0;
-        padding: 0 2em;
+        padding: 0.4em 2em;
         font-size: 1.4em;
         text-align: center;
     }
 
     :global(.body .bodyParagraph) {
-        margin: 1em 0;
-        padding: 0 2em;
+        padding: 1em 2em;
+        margin: 0;
     }
 
     :global(.body ul) {
-        margin: 1em 0;
-        padding: 0 2em 0 4em;
+        padding: 1em 2em 1em 4em;
     }
 
     :global(.body ol) {
-        margin: 1em 0;
-        padding: 0 2em 0 4em;
+        padding: 1em 2em 1em 4em;
     }
 
     :global(.body .bodyParagraph a) {
@@ -138,10 +120,14 @@
     }
 
     :global(.body .blogImg) {
-        margin: 1em 0;
-        padding: 0 2em;
+        padding: 1em 2em;
         display: flex;
         justify-content: center;
+        max-width: 90%;
+    }
+
+    :global(.body .innerImg) {
+        max-width: 100%;
     }
 
     :global(.body table) {
@@ -174,27 +160,34 @@
         background: var(--ddd);
         margin-bottom: 1em;
     }
-	
-	.teamAvatar {
-		vertical-align: middle;
-		border-radius: 50%;
-		height: 30px;
-		margin-right: 5px;
-		border: 0.25px solid #777;
-	}
 
-    .commentDivider {
-        margin: 1em 0 0;
-
+    .body {
+        position: relative;
+        max-height: 9em;
+        overflow: hidden;
     }
 
-    .authorAndDate {
-        color: var(--g999);
-        padding: 0 2em;
+    .fade {
+        position: absolute;
+        bottom: 0;
+        left: 0;
+        width: 100%;
     }
 
-    :global(.authorAndDate a) {
-        color: var(--g999);
+    .fadeTop {
+        height: 2em;
+        width: 100%;
+        background-image: linear-gradient(to bottom, var(--fffTransparent), var(--fff));
+    }
+
+    .fadeBottom {
+        height: 1em;
+        width: 100%;
+        background-color: var(--fff);
+    }
+
+    .viewFull {
+        padding: 0.2em 2em 1em;
     }
 </style>
 
@@ -208,27 +201,25 @@
         <div in:fly={{delay: duration, duration: duration, x: 150 * direction}} out:fly={{delay: 0, duration: duration, x: -150 * direction}} class="post">
             <h3>{title}</h3>
 
-            <div class="body">
+            <div class="body" bind:this={e} style="padding-bottom: {isOverflown ? '3em' : '0'}">
                 {#each body.content as paragraph}
                     {@html generateParagraph(paragraph)}
                 {/each}
+                {#if isOverflown}
+                    <div class="fade">
+                        <div class="fadeTop" />
+                        <div class="fadeBottom" />
+                    </div>
+                {/if}
+            </div>
+            
+            <div class="viewFull">
+                <a class="button" href="/blog/{id}">View Full Post</a>
             </div>
 
             <hr class="divider" />
 
-            <div class="authorAndDate">
-                <a href="/blog?filter={type}&page=1">{type}</a>
-                <img alt="author avatar" class="teamAvatar" src="{getAvatar(leagueTeamManagers, author)}" />
-                <span class="author">{@html getAuthor(leagueTeamManagers, author)} - </span>
-                <span class="date"><i>{parseDate(createdAt)}</i></span>
-            </div>
-
-            <!-- display comments -->
-            {#if !loadingComments && !home}
-                <hr class="divider commentDivider" />
-                <Comments {leagueTeamManagers} {comments} {total} postID={id} />
-            {/if}
-
+            <AuthorAndDate {type} {leagueTeamManagers} {author} {createdAt} />
         </div>
     {/key}
 {/if}
