@@ -9,13 +9,21 @@
 
 <style>
 	.playerAvatar {
+        position: relative;
 		vertical-align: middle;
-		height: 45px;
-		width: 45px;
+		height: 40px;
+		width: 40px;
 		background-position: center;
 		background-repeat: no-repeat;
 		background-size: auto 45px;
+        overflow: visible;
 	}
+    .teamLogo {
+        width: 21px;
+        position: absolute;
+        top: 0;
+        right: -16px;
+    }
 	.pos {
 		display: inline-flex;
 		align-items: center;
@@ -97,16 +105,12 @@
 		height: 45px !important;
 		overflow-y: hidden !important;
 	}
-	:global(.teamCell) {
-		max-width: 40px !important;
-    	padding: 0 5px;
-	}
 
 	:global(.nameCell) {
     	padding: 0 5px;
 	}
 
-	:global(span.nickname) {
+	span.nickname {
 		color: #888;
 		font-style: italic;
 		font-size: 0.8em;
@@ -136,21 +140,106 @@
 		color: #851313;
 	}
 
+    .additionalInfo {
+        display: inline-block;
+        font-size: 0.7em;
+    }
+
+    /* position text colors */
+	.text-QB {
+		color: var(--QB);
+	}
+
+	.text-WR {
+		color: var(--WR);
+	}
+
+	.text-RB {
+		color: var(--RB);
+	}
+
+	.text-TE {
+		color: var(--TE);
+	}
+
+	.text-K {
+		color: var(--K);
+	}
+
+	.text-DEF {
+		color: var(--DEF);
+	}
+
+	.text-DL {
+		color: var(--DL);
+	}
+
+	.text-LB {
+		color: var(--LB);
+	}
+
+	.text-DB {
+		color: var(--DB);
+	}
+
 	@media (max-width: 460px) {
 		:global(.nameCell) {
 			font-size: 0.9em;
 		}
 	}
 
+    @media (max-width: 355px) {
+        :global(.slot) {
+            width: 22px !important;
+            padding: 1px 2px 1px 2px !important;
+        }
+        :global(.avatar) {
+            padding: 0px 16px 0px 2px !important;
+        }
+        .pos {
+            display: inline-flex;
+            align-items: center;
+            justify-content: center;
+            border-radius: 8px;
+            width: 42px;
+            height: 24px;
+            font-size: 0.8em;
+        }
+		:global(.nameCell) {
+			font-size: 0.85em;
+		}
+    }
+
 </style>
 
 <Row>
 	<Cell class="slot playerCell"><span class="pos {playerSLotClass}">{playerSlot}</span></Cell>
 	{#if player.avatar}
-		<Cell class="avatar playerCell"><div class="playerAvatar" style="{player.avatar}" /></Cell>
+		<Cell class="avatar playerCell">
+            <div class="playerAvatar" style="{player.avatar}">
+                {#if player.team && player.poss != "DEF"}
+                    <img src="https://sleepercdn.com/images/team_logos/nfl/{player.team.toLowerCase()}.png" class="teamLogo" alt="team logo"/>
+                {/if}
+            </div>
+        </Cell>
 	{/if}
-	{#if player.poss != "DEF"}
-		<Cell class="playerCell teamCell">{player.team || ''}</Cell>
-	{/if}
-	<Cell class="playerCell nameCell" colspan={player.name != "Empty" ? (player.poss == "DEF" ? 2 : 1) : 3}>{@html player.name}</Cell>
+	<Cell class="playerCell nameCell" colspan={player.name != "Empty" ? 1 : 3}>
+        <div class="info">
+            <!-- name -->
+            {@html player.name}
+            <!-- name -->
+            {#if player.poss !== "DEF"}
+                <div class="additionalInfo">
+                    <span class="text-{player.poss}" >{@html player.poss}</span>
+                    {#if player.team}
+                        &nbsp;-&nbsp;
+                        {player.team}
+                    {/if}
+                </div>
+            {/if}
+        </div>
+        {#if player.nickname}
+            <span class="nickname">"{player.nickname}</span>
+        {/if}
+    </Cell>
 </Row>
