@@ -20,7 +20,6 @@ export const getLeagueStandings = async () => {
 
 	const yearData = leagueData.season;
 	const regularSeasonLength = leagueData.settings.playoff_week_start - 1;
-	const medianMatch = leagueData.settings.league_average_match == 1;
 	const divisions = leagueData.settings.divisions && leagueData.settings.divisions > 1;
     const rosters = rostersData.rosters;
 
@@ -46,7 +45,7 @@ export const getLeagueStandings = async () => {
         }
     }
 
-    if(divisions || medianMatch) {
+    if(divisions) {
         let week = 0;
         if(nflState.season_type == 'regular') {
             // max the week out at end of regular season
@@ -54,19 +53,19 @@ export const getLeagueStandings = async () => {
         } else if(nflState.season_type == 'post') {
             week = regularSeasonLength + 1;
         }
-    
+
         // if at least one week hasn't been completed, then standings can't be created
         if(week < 2) {
             return null;
         }
-    
+
         // pull in all matchup data for the season
         const matchupsPromises = [];
         for(let i = week - 1; i > 0; i--) {
             matchupsPromises.push(fetch(`https://api.sleeper.app/v1/league/${leagueID}/matchups/${i}`, {compress: true}))
         }
         const matchupsRes = await waitForAll(...matchupsPromises);
-    
+
         // convert the json matchup responses
         const matchupsJsonPromises = [];
         for(const matchupRes of matchupsRes) {
@@ -77,17 +76,16 @@ export const getLeagueStandings = async () => {
             }
         }
         const matchupsData = await waitForAll(...matchupsJsonPromises).catch((err) => { console.error(err); }).catch((err) => { console.error(err); });
-    
+
         // process all the matchups
         for(const matchup of matchupsData) {
-            standings = processStandings(matchup, standings, medianMatch, rosters);
+            standings = processStandings(matchup, standings, rosters);
         }
     }
 
 	const response = {
 		standingsInfo: standings,
 		yearData,
-		medianMatch,
 	}
 	
 	standingsStore.update(() => response);
@@ -95,9 +93,8 @@ export const getLeagueStandings = async () => {
 	return response;
 }
 
-const processStandings = (matchup, standingsData, medianMatch, rosters) => {
+const processStandings = (matchup, standingsData, rosters) => {
 	const matchups = {};
-	let scoresArray = [];
 	for(const match of matchup) {
 		if(!matchups[match.matchup_id]) {
 			matchups[match.matchup_id] = [];
@@ -109,17 +106,6 @@ const processStandings = (matchup, standingsData, medianMatch, rosters) => {
 			division: rosters[rosterID].settings.division,
 			points: match.points,
 		})
-
-		if(medianMatch == true){
-			scoresArray.push(match.points);
-		}
-	}
-	// calculating median score for the week
-	let medianScore;
-	if(medianMatch == true){
-		const numManagers = scoresArray.length;
-		scoresArray = scoresArray.sort((a, b) => b - a).slice(numManagers / 2 - 1, numManagers / 2 + 1);
-		medianScore = (scoresArray[0] + scoresArray[1]) / 2;
 	}
 
 	for(const matchupKey in matchups) {
@@ -127,19 +113,6 @@ const processStandings = (matchup, standingsData, medianMatch, rosters) => {
 		const teamB = matchups[matchupKey][1];
 	
 		const divisionMatchup = teamA.division && teamB.division && teamA.division == teamB.division;
-
-		// league average match
-		if(medianMatch == true) {
-			for(let i = 0; i < 2; i++) {
-				if(matchups[matchupKey][i].points > medianScore) {
-					standingsData[matchups[matchupKey][i].rosterID].wins ++;
-				} else if(matchups[matchupKey][i].points < medianScore) {
-					standingsData[matchups[matchupKey][i].rosterID].losses ++;
-				} else if(matchups[matchupKey][i].points == medianScore) {
-					standingsData[matchups[matchupKey][i].rosterID].ties ++;
-				}
-			}
-		}
 
         if(divisionMatchup) {
             if(teamA.points > teamB.points) {
