@@ -1,18 +1,16 @@
 
 import { enableBlog, getBlogPosts, getLeagueTeamManagers } from '$lib/utils/helper';
 
-export function load({ url, fetch }) {
+export function load({ fetch, params }) {
     if(!enableBlog) return false;
-
-    const queryPage = url?.searchParams?.get('page') || 1;
-    const filterKey = url?.searchParams?.get('filter') || '';
+    
+    const postID = params.slug;
     const postsData = getBlogPosts(fetch);
     const leagueTeamManagersData = getLeagueTeamManagers();
 
     return {
-        queryPage,
         postsData,
-        filterKey,
+        postID,
         leagueTeamManagersData,
     };
 }
